feat(TA-1034): add partitioned cookie configuration (#94)
## Problem

* [TA-1034](https://techfromsage.atlassian.net/browse/TA-1034)

On analysing Elevate's use of LTI.js, we identified that it requires
third party cookies. The library has not yet been updated to change this
particular feature.

In order for LTI launches to continue to work in Elevate (at least for
Chrome-based browsers), we require a solution that allows these cookies
to be processed in a PKCE workflow.

The current solution to this issue is [partioned
cookies](https://developer.mozilla.org/en-US/docs/Web/Privacy/Privacy_sandbox/Partitioned_cookies).

Elevate cannot support this at an LTI.js level, nor at the Express.js
level so we must resort to a supporting configuration in nginx.

This new property will load a supporting nginx configuration. 

[TA-1034]:
https://techfromsage.atlassian.net/browse/TA-1034?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ